Move user state client django app into a less ambiguosly named django app
